568Div2
Problem Statement
    
Two positive integers A and B are called similar if A <= 10*B and B <= 10*A. For example, 1 and 10 are similar, but 1 and 11 are not.  You are given ints lower and upper. You must select as many integers as possible so that:
each selected integer is between lower and upper, inclusive;
no two selected integers are similar.
Return the maximum number of selected integers.