582 Div2
Problem Statement:  
Magical Girl Iris loves perfect squares. A positive integer n is a perfect square if and only if there is a positive integer b >= 1 such that b*b = n. For example, 1 (=1*1), 16 (=4*4), and 169 (=13*13) are perfect squares, while 2, 54, and 48 are not. Iris also likes semi-squares. A positive integer n is called a semi-square if and only if there are positive integers a >= 1 and b > 1 such that a < b and a*b*b = n. For example, 81 (=1*9*9) and 48 (=3*4*4) are semi-squares, while 24, 63, and 125 are not. (Note that we require that a < b. Even though 24 can be written as 6*2*2, that does not make it a semi-square.) You are given a int N. Return "Yes" (quotes for clarity) if N is a semi-square number. Otherwise, return "No".