534 Div2
Problem Statement:
Elly wants to write a program that lists all the files in a given directory. She already has the list of all the files. You will be given this list as a String[] files. In addition to the names of files, the variable files will contain exactly two additional elements: the current directory (the String "."), and the parent directory (the String ".."). These two elements may be anywhere in files. However, Elly wants them to be the last two elements. In order to move the two directories to the last two positions in files, she wants you to perform the following steps:
If "." and ".." are the last two elements of files (in any order), you are done.
Find the first element of files that is either "." or "..". Swap it with the last element of files.
If "." and ".." are now the last two elements of files (in any order), you are done.
Find the first element of files that is either "." or "..". Swap it with the element of files that is one position before the last one.
Your method must perform the above steps and return a String[] containing the modified order of elements in files.